toolkit/tools/messaging_matrix/execution/requirements.txt
openai>=1.12.0
python-docx>=1.1.0
